extract partial structures to another file
