+++ src/record/mod.rs
@@ -1,10 +1,12 @@
-use bytes::{BufMut, Bytes, BytesMut};
+pub(crate) mod partially_serialized;
 
-use crate::{blob::File, error::ValidationErrorKind, prelude::*};
+use bytes::{BufMut, Bytes, BytesMut};
+pub(crate) use partially_serialized::PartiallySerializedRecord;
+
+use crate::{error::ValidationErrorKind, prelude::*};
 
 pub(crate) const RECORD_MAGIC_BYTE: u64 = 0xacdc_bcde;
 const DELETE_FLAG: u8 = 0x01;
-const MAX_SINGLE_PASS_DATA_SIZE: usize = 4 * 1024;
 
 #[derive(Serialize, Deserialize, Default, Clone, PartialEq)]
 pub struct Record {
@@ -117,109 +119,6 @@ impl Meta {
     }
 }
 
-enum PartiallySerializedData {
-    SinglePass(BytesMut),
-    DoublePass(BytesMut, Bytes),
-}
-
-impl PartiallySerializedData {
-    pub(crate) async fn write_to_file(self, file: &File) -> Result<u64> {
-        match self {
-            PartiallySerializedData::SinglePass(b) => {
-                Self::write_single_pass(b.freeze(), file).await
-            }
-            PartiallySerializedData::DoublePass(h, d) => {
-                Self::write_double_pass(h.freeze(), d, file).await
-            }
-        }
-    }
-
-    async fn write_single_pass(buf: Bytes, file: &File) -> Result<u64> {
-        let len = buf.len() as u64;
-        Self::process_file_result(file.write_append_all(buf).await).map(|_| len)
-    }
-
-    async fn write_double_pass(head: Bytes, data: Bytes, file: &File) -> Result<u64> {
-        let len = (head.len() + data.len()) as u64;
-        Self::process_file_result(file.write_append_all_buffers(head, data).await).map(|_| len)
-    }
-
-    fn process_file_result(result: std::result::Result<(), std::io::Error>) -> Result<()> {
-        result.map_err(|e| -> anyhow::Error {
-            match e.kind() {
-                kind if kind == IOErrorKind::Other || kind == IOErrorKind::NotFound => {
-                    Error::file_unavailable(kind).into()
-                }
-                _ => e.into(),
-            }
-        })
-    }
-}
-
-pub(crate) struct PartiallySerializedRecord {
-    data: PartiallySerializedData,
-    offset_pos: usize,
-    header_checksum_pos: usize,
-    header_start: usize,
-    header_len: usize,
-}
-
-impl PartiallySerializedRecord {
-    fn new(
-        data: PartiallySerializedData,
-        offset_pos: usize,
-        header_checksum_pos: usize,
-        header_start: usize,
-        header_len: usize,
-    ) -> Self {
-        Self {
-            data,
-            offset_pos,
-            header_checksum_pos,
-            header_start,
-            header_len,
-        }
-    }
-
-    fn into_serialized_with_header_checksum(
-        self,
-        blob_offset: u64,
-    ) -> Result<(PartiallySerializedData, u32)> {
-        let Self {
-            mut data,
-            offset_pos,
-            header_checksum_pos,
-            header_start,
-            header_len,
-        } = self;
-        let checksum: u32;
-        match data {
-            PartiallySerializedData::SinglePass(ref mut buf) => {
-                let offset_slice = &mut buf[offset_pos..];
-                bincode::serialize_into(offset_slice, &blob_offset)?;
-                let header_slice = &buf[header_start..(header_start + header_len)];
-                checksum = CRC32C.checksum(header_slice);
-                let checksum_slice = &mut buf[header_checksum_pos..];
-                bincode::serialize_into(checksum_slice, &checksum)?;
-            }
-            PartiallySerializedData::DoublePass(ref mut buf, _) => {
-                let offset_slice = &mut buf[offset_pos..];
-                bincode::serialize_into(offset_slice, &blob_offset)?;
-                let header_slice = &buf[header_start..(header_start + header_len)];
-                checksum = CRC32C.checksum(header_slice);
-                let checksum_slice = &mut buf[header_checksum_pos..];
-                bincode::serialize_into(checksum_slice, &checksum)?;
-            }
-        }
-        Ok((data, checksum))
-    }
-
-    pub async fn write_to_file(self, file: &File, offset: u64) -> Result<(u64, u32)> {
-        let (data, checksum) = self.into_serialized_with_header_checksum(offset)?;
-        data.write_to_file(file).await.map(|l| (l, checksum))
-    }
-}
-
 impl Record {
     pub(crate) fn new(header: Header, meta: Meta, data: Bytes) -> Self {
         Self { header, meta, data }
@@ -259,14 +158,15 @@ impl Record {
         let header_len = head.len();
         self.meta.to_raw_into((&mut head).writer())?;
         let Self { data, header, .. } = self;
-        let data = if head.len() + data.len() > MAX_SINGLE_PASS_DATA_SIZE {
-            PartiallySerializedData::DoublePass(head, data)
-        } else {
-            head.extend_from_slice(&data);
-            PartiallySerializedData::SinglePass(head)
-        };
         Ok((
-            PartiallySerializedRecord::new(data, header_len - 24, header_len - 4, 0, header_len),
+            PartiallySerializedRecord::new(
+                head,
+                data,
+                header_len - 24,
+                header_len - 4,
+                0,
+                header_len,
+            ),
             header,
         ))
     }
